+++ docs/components/content/CodeWindow.tsx
