+++ docs/components/content/Quote.tsx
