+++ docs/pages/for-developers.tsx
@@ -4,18 +4,18 @@ import Image from 'next/image';
 import Link from 'next/link';
 
 import { useMediaQuery } from '../lib/media';
-import { IntroWrapper, IntroHeading, IntroLead } from '../components/marketing/Intro';
-import { CommunityCta } from '../components/marketing/CommunityCta';
+import { IntroWrapper, IntroHeading, IntroLead } from '../components/content/Intro';
+import { CommunityCta } from '../components/content/CommunityCta';
 import { FrontEndLogos } from '../components/icons/FrontEndLogos';
 import { Highlight } from '../components/primitives/Highlight';
 import { Relational } from '../components/icons/Relational';
-import { TweetBox } from '../components/marketing/TweetBox';
-import { MWrapper } from '../components/marketing/MWrapper';
+import { TweetBox } from '../components/content/TweetBox';
+import { MWrapper } from '../components/content/MWrapper';
 import { Typescript } from '../components/icons/Typescript';
-import { Section } from '../components/marketing/Section';
-import { CodeBox } from '../components/marketing/CodeBox';
+import { Section } from '../components/content/Section';
+import { CodeBox } from '../components/content/CodeBox';
 import { Button } from '../components/primitives/Button';
-import { EndCta } from '../components/marketing/EndCta';
+import { EndCta } from '../components/content/EndCta';
 import { Postgres } from '../components/icons/Postgres';
 import { Emoji } from '../components/primitives/Emoji';
 import { GraphQl } from '../components/icons/GraphQl';
@@ -24,7 +24,7 @@ import { ArrowR } from '../components/icons/ArrowR';
 import { Nextjs } from '../components/icons/Nextjs';
 import { Shield } from '../components/icons/Shield';
 import { Prisma } from '../components/icons/Prisma';
-import { Pill } from '../components/marketing/Pill';
+import { Pill } from '../components/content/Pill';
 import { Tick } from '../components/icons/Tick';
 import { Cli } from '../components/icons/Cli';
 import { Page } from '../components/Page';
