+++ docs/pages/for-organisations.tsx
@@ -4,18 +4,18 @@ import Image from 'next/image';
 import Link from 'next/link';
 
 import { useMediaQuery } from '../lib/media';
-import { IntroWrapper, IntroHeading, IntroLead } from '../components/marketing/Intro';
+import { IntroWrapper, IntroHeading, IntroLead } from '../components/content/Intro';
 import { Highlight } from '../components/primitives/Highlight';
 import { ClientLogos } from '../components/icons/ClientLogos';
-import { TweetBox } from '../components/marketing/TweetBox';
-import { MWrapper } from '../components/marketing/MWrapper';
-import { Section } from '../components/marketing/Section';
+import { TweetBox } from '../components/content/TweetBox';
+import { MWrapper } from '../components/content/MWrapper';
+import { Section } from '../components/content/Section';
 import { Thinkmill } from '../components/icons/Thinkmill';
 import { Button } from '../components/primitives/Button';
-import { Quote } from '../components/marketing/Quote';
+import { Quote } from '../components/content/Quote';
 import { Type } from '../components/primitives/Type';
 import { ArrowR } from '../components/icons/ArrowR';
-import { Pill } from '../components/marketing/Pill';
+import { Pill } from '../components/content/Pill';
 import { Tick } from '../components/icons/Tick';
 import { Page } from '../components/Page';
