+++ docs/pages/why-keystone.tsx
@@ -7,23 +7,23 @@ import { useMediaQuery } from '../lib/media';
 import { Highlight } from '../components/primitives/Highlight';
 import { Relationship } from '../components/icons/Relationship';
 import { WhyKeystone } from '../components/icons/WhyKeystone';
-import { MWrapper } from '../components/marketing/MWrapper';
-import { TweetBox } from '../components/marketing/TweetBox';
+import { MWrapper } from '../components/content/MWrapper';
+import { TweetBox } from '../components/content/TweetBox';
 import { Typescript } from '../components/icons/Typescript';
 import { Automated } from '../components/icons/Automated';
 import { Migration } from '../components/icons/Migration';
-import { Section } from '../components/marketing/Section';
+import { Section } from '../components/content/Section';
 import { Button } from '../components/primitives/Button';
-import { EndCta } from '../components/marketing/EndCta';
+import { EndCta } from '../components/content/EndCta';
 import { Emoji } from '../components/primitives/Emoji';
 import { Updates } from '../components/icons/Updates';
-import { Quote } from '../components/marketing/Quote';
+import { Quote } from '../components/content/Quote';
 import { Type } from '../components/primitives/Type';
 import { ArrowR } from '../components/icons/ArrowR';
 import { Custom } from '../components/icons/Custom';
 import { Editor } from '../components/icons/Editor';
 import { Filter } from '../components/icons/Filter';
-import { Pill } from '../components/marketing/Pill';
+import { Pill } from '../components/content/Pill';
 import { Shield } from '../components/icons/Shield';
 import { Watch } from '../components/icons/Watch';
 import { Cli } from '../components/icons/Cli';
@@ -200,7 +200,7 @@ export default function WhyKeystonePage() {
                 Extend the CRUD API for more control over what you do. Customize it to your frontend
                 needs.
               </Type>
-              <animateMotion
+              <a
                 href="https://github.com/keystonejs/keystone/tree/master/examples/extend-graphql-schema"
                 target="_blank"
                 rel="noopener noreferrer"
